requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gspread==5.11.3
oauth2client==4.1.3
python-dotenv==1.0.0
loguru==0.7.2
openai>=1.0.0